requirements.txt
torch==2.0.1
torchvision==0.15.2
tensorflow==2.6.0
flask==2.3.2
pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
protobuf==3.20.3
